+++ click_extended/decorators/__init__.py
@@ -1,9 +1,13 @@
 """Initialization file for the `click_extended.decorators` module."""
 
+from click_extended.decorators.random_bool import random_bool
+from click_extended.decorators.random_float import random_float
 from click_extended.decorators.random_integer import random_integer
 from click_extended.decorators.random_string import random_string
 
 __all__ = [
+    "random_bool",
+    "random_float",
     "random_integer",
     "random_string",
 ]